requirements.txt
# LangChain相关依赖
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-chroma>=0.1.0

# LangGraph工作流框架
langgraph>=0.1.0

# DOCX文件处理
python-docx==1.2.0

# 数据模型和验证
pydantic>=2.0.0

# 类型扩展支持
typing-extensions>=4.0.0
